Fix user-view-intergrate - #39
