+++ board/views/user.py
@@ -1,26 +1,11 @@
 from ..responses import *
-from ..utils import send_json, pop_args
+from ..utils import send_json, pop_args, byte_to_dict
 from ..models import User
 from django.contrib.auth.hashers import make_password
 from django.views import View
-from urllib import parse
 import sys
 sys.path.append("../../")
 from DjangoCRUDBoard import settings
-
-def byte_to_dict(data):
-    body_unicode = data.decode('utf-8')
-    body = body_unicode.replace("&", "=")
-    body_list = body.split("=")
-    body_key = []
-    body_value = []
-    for i in range(len(body_list)):
-        if i % 2 == 0:
-            body_key.append(body_list[i])
-        else:
-            body_value.append(parse.unquote(body_list[i]))
-    dict_body = dict(zip(body_key, body_value))
-    return dict_body
 
 class UserView(View):
     def post(self, request):
@@ -30,11 +15,9 @@ class UserView(View):
             return send_json(illegalArgument)
         filtered = User.objects.filter(username=dic['username'])
         if filtered.count() != 0:
-            data = userAlreadyRegistered
-        else:
-            User.objects.create_user(*dic.values())
-            data = registerSucceed
-        return send_json(data)
+            return send_json(userAlreadyRegistered)
+        User.objects.create_user(*dic.values())
+        return send_json(registerSucceed)
 
     def delete(self, request):
         keys = ['username', 'email', 'password']
@@ -44,14 +27,11 @@ class UserView(View):
             return send_json(illegalArgument)
         filtered = User.objects.filter(username=dic['username'], email=dic['email'])
         if filtered.count() == 0:
-            data = noUser 
-        else:
-            if filtered[0].check_password(dic['password']):
-                filtered.delete()
-                data = deleteUserSucceed
-            else:
-                data = userDoesNotMatch
-        return send_json(data)
+            return send_json(noUser)
+        if not filtered[0].check_password(dic['password']):
+            return send_json(userDoesNotMatch)
+        filtered.delete()
+        return send_json(deleteUserSucceed)
 
     def put(self, request):
         keys = ['username', 'email', 'password', 'm_username', 'm_email', 'm_password']
@@ -63,27 +43,28 @@ class UserView(View):
             return send_json(illegalModifyArgument)
         filtered = User.objects.filter(username=dic['username'], email=dic['email'])
         if filtered.count() == 0:
-            data = noUser 
-        else:
-            if filtered[0].check_password(dic['password']): 
-                # 변경하려는 아이디, 이메일, 패스워드가 기존과 같다면 어떻게 처리해줘야할지..
-                update_value = dic.copy()
-                del update_value['m_username'], update_value['m_email'], update_value['m_password']
-                '''
-                해당 딕셔너리 copy 해준 후 뒷 부분 없애줌. 각 if문 마다 filtered.update해주기에는 비효율적이어서 
-                이와 같이 기존 값을 그대로 가지고 있게 하여 m_username만 있고 m_email, m_password없는 등 부분적으로
-                컬럼이 없을시에도 기존 값을 가지고 있음으로써 1번만 update해주면 됨
-                '''
-                if dic['m_username']: 
-                    update_value['username'] = dic['m_username']
-                if dic['m_email']:
-                    update_value['email'] = dic['m_email']
-                if dic['m_password']:
-                    update_value['password'] = dic['m_password']
+            return send_json(noUser)
+        if not filtered[0].check_password(dic['password']):
+            return send_json(userDoesNotMatch)
+        # 변경하려는 username이 기존에 존재하는지 처리 분기문
+        filtered_exist = User.objects.filter(username=dic['m_username'])
+        if filtered_exist.count() != 0:
+            return send_json(userAlreadyRegistered)
+        # 변경하려는 아이디, 이메일, 패스워드가 기존과 같다면 어떻게 처리해줘야할지..
+        update_value = dic.copy()
+        del update_value['m_username'], update_value['m_email'], update_value['m_password']
+        '''
+        해당 딕셔너리 copy 해준 후 뒷 부분 없애줌. 각 if문 마다 filtered.update해주기에는 비효율적이어서 
+        이와 같이 기존 값을 그대로 가지고 있게 하여 m_username만 있고 m_email, m_password없는 등 부분적으로
+        컬럼이 없을시에도 기존 값을 가지고 있음으로써 1번만 update해주면 됨
+        '''
+        if dic['m_username']: 
+            update_value['username'] = dic['m_username']
+        if dic['m_email']:
+            update_value['email'] = dic['m_email']
+        if dic['m_password']:
+            update_value['password'] = dic['m_password']
 
-                filtered.update(username=update_value['username'], email=update_value['email'], password=make_password(update_value['password']))
-                # 현재 이방식은 normalize_email, normalize_username 를 안해주는데 괜찮을지?! create_user 함수 확인 요구됨. 
-                data = modifyUserSucceed
-            else:
-                data = userDoesNotMatch
-        return send_json(data)
+        filtered.update(username=update_value['username'], email=update_value['email'], password=make_password(update_value['password']))
+        # 현재 이방식은 normalize_email, normalize_username 를 안해주는데 괜찮을지?! create_user 함수 확인 요구됨. 
+        return send_json(modifyUserSucceed)
